+++ src/envoy/extensions/wasm/BUILD
@@ -25,6 +25,9 @@ envoy_cc_library(
     srcs = [
         "wasm.cc",
     ],
+    hdrs = [
+        "context.h",
+    ],
     repository = "@envoy",
     visibility = ["//visibility:public"],
     deps = [
