+++ src/envoy/extensions/wasm/context.h
@@ -1,0 +1,38 @@
+/* Copyright 2018 Istio Authors. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+namespace Envoy {
+namespace Extensions {
+namespace Common {
+namespace Wasm {
+namespace Istio {
+
+class IstioContext : public Context {
+ public:
+  IstioContext() : Context() {}
+  IstioContext(Wasm* wasm) : Context(wasm) {}
+  IstioContext(Wasm* wasm, const PluginSharedPtr& plugin)
+      : Context(wasm, plugin) {}
+  IstioContext(Wasm* wasm, uint32_t root_context_id,
+               const PluginSharedPtr& plugin)
+      : Context(wasm, root_context_id, plugin) {}
+  ~IstioContext() = default;
+};
+}  // namespace Istio
+}  // namespace Wasm
+}  // namespace Common
+}  // namespace Extensions
+}  // namespace Envoy
